refactor(fmt): split main.rs into per-responsibility modules
Pure move: every line of the 2349-line `main.rs` lands verbatim in one of
15 modules (`cli`, `paths`, `oxfmt`, `run`, `status`, `options`,
`tailwind_sort`, `stdin`, `walk`, `svelte_pipeline`, `output`,
`native_js`, `native_json`, `native_css`), leaving `main.rs` with the
crate doc, the allocator, the `mod` list and `fn main`. Only `use` lines
and visibility (`pub(crate)`) were adjusted; two intra-doc links were
path-qualified so `cargo doc` still resolves them across modules.
